Switch to a swipeable tab view. Update the text in the Info tab to reflect all the changes I've made so far.
There's a remaining bug I need to chase down -- when you change the device orientation (eg, portrait to landscape), the tabbed display jumps to a random tab. This is a fairly common problem with SwiftUI apps, according to StackOverflow. I'll figure this out next but wanted to get the current changes committed.