Added copyright and BSD license text to file headers where missing.
Get the text on the info page correct.

UI cleanup -- make sure that the displays all fit on the available space on all the different phone sizes in all orientations. I had to use a smaller font and reduce some image sizes in the TimeView to make that work, but now you can see everything without overlapping with the tab indicator that SwiftUI puts at the bottom of the pages.